requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
cffi==1.15.1
cryptography==38.0.2
mysql-connector-python==8.0.31
protobuf==3.20.1
pycparser==2.21
pyperclip==1.8.2
PySimpleGUI==4.60.4
